+++ app/components/Analyst/Assessments/AssessmentsForm.tsx
@@ -65,11 +65,11 @@ const AssessmentsForm: React.FC<Props> = ({
             _jsonData: e.formData,
             _assessmentType: slug,
           },
+          connections: [],
         },
         onCompleted: () => {
           setIsFormSaved(true);
           if (
-            slug === 'screening' &&
             e.formData?.nextStep === 'Needs 2nd review' &&
             e.formData?.nextStep !== formData?.nextStep
           ) {
@@ -82,6 +82,7 @@ const AssessmentsForm: React.FC<Props> = ({
                 applicationId: queryFragment.rowId,
                 host: window.location.origin,
                 ccbcNumber: addedContext?.ccbcNumber,
+                assessmentType: slug,
               }),
             }).then((response) => {
               if (!response.ok) {