+++ app/schema/mutations/assessment/createAssessment.ts
@@ -3,13 +3,23 @@ import { createAssessmentMutation } from '__generated__/createAssessmentMutation
 import useMutationWithErrorMessage from '../useMutationWithErrorMessage';
 
 const mutation = graphql`
-  mutation createAssessmentMutation($input: CreateAssessmentFormInput!) {
+  mutation createAssessmentMutation(
+    $input: CreateAssessmentFormInput!
+    $connections: [ID!]!
+  ) {
     createAssessmentForm(input: $input) {
-      assessmentData {
+      assessmentData
+        @prependNode(
+          connections: $connections
+          edgeTypeName: "AssessmentDataEdge"
+        ) {
         id
         rowId
         jsonData
         createdAt
+        updatedAt
+        updatedBy
+        assessmentDataType
       }
     }
   }