+++ app/tests/pages/analyst/application/[applicationId]/assessments/screening.test.tsx
@@ -170,6 +170,7 @@ describe('The index page', () => {
         },
         _assessmentType: 'screening',
       },
+      connections: [],
     });
   });
 
@@ -192,6 +193,7 @@ describe('The index page', () => {
         },
         _assessmentType: 'screening',
       },
+      connections: [],
     });
 
     await act(async () => {